Full Redo of portfolio following notes.
NOTES: ok here's his notes
sure. Let me write my notes here in the meantime.
Overall, this portfolio doesn't give me a good understanding of anything he's done in his projects - and it's not great to have what seems to be many unfinished projects or basic school assignments. Block Run is an example of something I'd never show to an employer or hiring manager. "Prototype" in Godot is a good example of just poor naming and not showcasing the actual that was done - which was generating terrain via script. Instead of talking about how he couldn't get much done, he should focus on what WAS done and make it clear how it worked. None of his projects have code snippets or any kind of diagram breaking down his work process, which is not great. Some of them have git repots which is cool, but I wouldn't click on any of them unless I saw code snippets or better breakdowns of the systems he worked on on the portfolio page itself.
As an excercise he should focus on 3 main projects, based on what he feels is important and demonstrates the best work he's done, and then break them down or showcase some of the inner workings or thought process (via flow charts etc) on those pages. If a person who does not know him, reads the page - they should have a good understanding of the work he's done and why it's important for the game
as a tech artist, I'd expect to see tools for Blender or Maya, and a lot more python work based on his skillset
He also has no resume on his website